interpretation: extract the shared data types into types.rs
Per Nico's #883 review: give the module's cross-cutting data types one home
instead of scattering them across mod.rs / graph. New `interpretation/types.rs`
holds the vocabulary that flows between submodules:
- ProposedInstance   (was mod.rs)      — the parse output
- InstanceContext + ExistingInstances  (was graph/mod.rs) — the existing-graph snapshot
- InterpretationOp   (was graph/write.rs) — a planned write

Chose the module level (not graph/types.rs): none of these is graph-internal —
ProposedInstance is `parse`'s output, InterpretationOp flows write→run,
InstanceContext/ExistingInstances flow read→prompt→dedup. Consumers reach them
via the existing `pub use` re-export, so call sites are unchanged; submodules
that referenced them through `graph` now import from `types` directly.

Shrinks graph/write.rs (789→759) and empties the type defs out of mod.rs
(now 52 lines) and graph/mod.rs (92, helpers only). Pure move, no logic
changes. All 45 interpretation tests pass (--test-threads=1); no new warnings.
