+++ rust-executor/src/perspectives/interpretation/graph/read.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::perspectives::interpretation::types::{ExistingInstances, InstanceContext};
 use crate::perspectives::model_query::types::{ModelShape, ParentScope};
 use crate::perspectives::perspective_instance::PerspectiveInstance;
 use std::collections::{BTreeMap, HashMap};
